+++ app/src/main/java/com/space/quizapp/domain/repository/UserAuthRepository.kt
@@ -2,13 +2,10 @@ package com.space.quizapp.domain.repository
 
 import com.space.quizapp.domain.model.UserDomainModel
 
-interface UserRepository {
+interface UserAuthRepository {
     //Room
     suspend fun getUserByUsername(username: String): UserDomainModel?
     suspend fun saveUser(user: UserDomainModel)
     suspend fun getUserPoints(username: String): Double?
-    //DataStore
-    suspend fun getUserSession(): Result<String>
-    suspend fun saveUserSession(user: String)
-    suspend fun clearUserSession()
+
 }
