+++ app/src/main/java/com/space/quizapp/domain/usecase/current_user/clear/ClearUserSessionUseCaseImpl.kt
@@ -1,8 +1,8 @@
 package com.space.quizapp.domain.usecase.current_user.clear
 
-import com.space.quizapp.domain.repository.UserRepository
+import com.space.quizapp.domain.repository.UserSessionRepository
 
-class ClearUserSessionUseCaseImpl(private val userRepository: UserRepository): ClearUserSessionUseCase {
+class ClearUserSessionUseCaseImpl(private val userRepository: UserSessionRepository): ClearUserSessionUseCase {
 
     override suspend fun invoke() {
         userRepository.clearUserSession()
