+++ app/src/main/java/com/space/quizapp/domain/usecase/current_user/get/GetUserSessionUseCaseImpl.kt
@@ -1,8 +1,8 @@
 package com.space.quizapp.domain.usecase.current_user.get
 
-import com.space.quizapp.domain.repository.UserRepository
+import com.space.quizapp.domain.repository.UserSessionRepository
 
-class GetUserSessionUseCaseImpl(private val userRepository: UserRepository):GetUserSessionUseCase{
+class GetUserSessionUseCaseImpl(private val userRepository: UserSessionRepository):GetUserSessionUseCase{
     override suspend fun invoke(): Result<String> {
         return userRepository.getUserSession()
     }
