+++ app/src/main/java/com/space/quizapp/domain/usecase/current_user/save/SaveUserSessionUseCaseImpl.kt
@@ -1,8 +1,8 @@
 package com.space.quizapp.domain.usecase.current_user.save
 
-import com.space.quizapp.domain.repository.UserRepository
+import com.space.quizapp.domain.repository.UserSessionRepository
 
-class SaveUserSessionUseCaseImpl(private val userRepository: UserRepository) :
+class SaveUserSessionUseCaseImpl(private val userRepository: UserSessionRepository) :
     SaveUserSessionUseCase {
     override suspend fun saveUserSession(username: String) {
         userRepository.saveUserSession(username)
