+++ app/src/main/res/values/strings.xml
@@ -20,5 +20,5 @@
     <string name="yes">კი</string>
     <string name="no">არა</string>
     <string name="gpa_score">GPA -&#160;</string>
-    <string name="hello_user">გამარჯობა, </string>
+    <string name="hello_user">გამარჯობა,&#160; </string>
 </resources>
